자신의 글만 삭제 가능

+++ src/main/java/com/prj2spring/controller/board/BoardController.java
@@ -3,6 +3,7 @@ package com.prj2spring.controller.board;
 import com.prj2spring.domain.board.Board;
 import com.prj2spring.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -49,8 +50,15 @@ public class BoardController {
     }
 
     @DeleteMapping("{id}")
-    public void delete(@PathVariable Integer id) {
-        service.remove(id);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Object> delete(@PathVariable Integer id,
+                                         Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     }
 
     @PutMapping("edit")
